requirements.txt
moviepy==1.0.3
google-auth-oauthlib==1.2.0
google-api-python-client==2.108.0
openai==1.3.0
pillow==10.1.0
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
